Move graphics and ecs
