+++ Light Vox Engine/ECS/Entity.h
@@ -1,5 +1,5 @@
 #pragma once 
-#include "stdafx.h"
+#include "../stdafx.h"
 
 class Entity
 {
