+++ Light Vox Engine/Graphics/FrameResource.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "stdafx.h"		
+#include "../stdafx.h"		
 #include "ShaderDefinitions.h"
 
 class Camera;
