+++ Light Vox Engine/Graphics/GraphicsCore.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "stdafx.h"
+#include "../stdafx.h"
 
 class Camera;
 class FrameResource;
