+++ Light Vox Engine/Light Vox Engine.vcxproj
@@ -234,40 +234,39 @@ for /r "$(SolutionDir)_Binary" %%x in (*.cso) do move "%%x" "$(SolutionDir)_Bina
     </FxCompile>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="Camera.cpp" />
-    <ClCompile Include="ComponentManager.cpp" />
+    <ClCompile Include="ECS\ComponentManager.cpp" />
+    <ClCompile Include="ECS\EntityManager.cpp" />
     <ClCompile Include="Engine.cpp" />
-    <ClCompile Include="EntityManager.cpp" />
-    <ClCompile Include="FrameResource.cpp" />
-    <ClCompile Include="GraphicsCore.cpp" />
+    <ClCompile Include="Graphics\Camera.cpp" />
+    <ClCompile Include="Graphics\FrameResource.cpp" />
+    <ClCompile Include="Graphics\GraphicsCore.cpp" />
+    <ClCompile Include="Graphics\ObjLoader.cpp" />
     <ClCompile Include="JobSystem\JobManager.cpp" />
     <ClCompile Include="JobSystem\JobSequence.cpp" />
     <ClCompile Include="Main.cpp" />
-    <ClCompile Include="ObjLoader.cpp" />
     <ClCompile Include="GameTime.cpp" />
     <ClCompile Include="Physics\Physics.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Camera.h" />
-    <ClInclude Include="ComponentManager.h" />
     <ClInclude Include="Configs.h" />
     <ClInclude Include="d3dx12.h" />
     <ClInclude Include="DXSampleHelper.h" />
+    <ClInclude Include="ECS\ComponentManager.h" />
+    <ClInclude Include="ECS\Entity.h" />
+    <ClInclude Include="ECS\EntityComponents.h" />
+    <ClInclude Include="ECS\EntityManager.h" />
     <ClInclude Include="Engine.h" />
-    <ClInclude Include="Entity.h" />
-    <ClInclude Include="EntityComponents.h" />
-    <ClInclude Include="EntityManager.h" />
-    <ClInclude Include="FrameResource.h" />
-    <ClInclude Include="GraphicsCore.h" />
-    <ClInclude Include="JobSystem\ComponentManager.h" />
+    <ClInclude Include="Graphics\Camera.h" />
+    <ClInclude Include="Graphics\FrameResource.h" />
+    <ClInclude Include="Graphics\GraphicsCore.h" />
+    <ClInclude Include="Graphics\ObjLoader.h" />
+    <ClInclude Include="Graphics\ShaderDefinitions.h" />
     <ClInclude Include="JobSystem\ConcurrentQueue.h" />
     <ClInclude Include="JobSystem\CpuJob.h" />
     <ClInclude Include="JobSystem\JobManager.h" />
     <ClInclude Include="JobSystem\JobSequence.h" />
     <ClInclude Include="LVutils.h" />
-    <ClInclude Include="ObjLoader.h" />
     <ClInclude Include="Physics\Physics.h" />
-    <ClInclude Include="ShaderDefinitions.h" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="GameTime.h" />
   </ItemGroup>
